refactoring ui package
