+++ app/src/main/java/com/example/testapplication/StartApp.kt
@@ -2,6 +2,7 @@ package com.example.testapplication
 
 import android.app.Application
 import com.example.testapplication.di.catRepositoryModule
+import com.example.testapplication.di.localModule
 import com.example.testapplication.di.mainViewModelModule
 import com.example.testapplication.di.networkModule
 import org.koin.core.context.startKoin
@@ -13,6 +14,7 @@ class StartApp: Application() {
         startKoin {
             modules(listOf(
                 networkModule,
+                localModule,
                 catRepositoryModule,
                 mainViewModelModule
             ))
